otf2: update homepage, livecheck
The existing homepage URL for `otf2` redirects to an overview page
and this page no longer contains a tarball link, so livecheck is
giving an `Unable to get versions` error. This updates the homepage
to resolve the redirection and updates the `livecheck` block URL to
check the upstream download page, which links to the latest tarball.